requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-socketio
eventlet
gunicorn
psycopg2-binary
python-dotenv
qrcode